Try to load libraries directly on Linux
Fedora Packaging Guidelines for Java require the use of
System.load() instead of System.loadLibrary(). Also, the
name of the library is fixed. So the whole process of
guessing the architecture and path can be sidestepped.

Actual atlas library that is loaded is determined at runtime,
based what is installed. Thus package dependencies determine
the outcome, and there's no need to check the architecture.

Unfortunately there doesn't seem to be a way to depend
on any version of the atlas-*-devel headers, so during
build, atlas-devel is required. On installation however,
uninstalling atlas and installing e.g. atlas-sse3 will give
a nice speedup.